Rename threading_methods file to numba_methods
